Move args to options
